Moving some classes to their own files, increase to retry delay for v4 queries
