+++ GQLScanning.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol GQLScanning: GQLElement {
+    func scan(query: GQLQuery, pageData: Any, parent: GQLNode?, level: Int) -> [GQLQuery]
+}
